RAMA_agentic_app/requirements.txt
gradio==3.50.2
openai==1.75.0
psycopg2-binary>=2.9.6
python-dotenv
pydantic<2.0.0
fastapi
uvicorn
secure-smtplib

